git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/pario2@19822 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ test/extern/sungeun/array_ops.h
@@ -1,2 +1,4 @@
-void clear_int_array(int *A, int len);
-void copy_int_array(int *dest, int *src, int len);
+#include <stdint.h>
+
+void clear_int_array(int64_t *A, int len);
+void copy_int_array(int64_t *dest, int64_t *src, int len);
